requirements.txt
python-ldap==2.4.19
requests==2.5.3
wsgiref==0.1.2
